CLEANUP: ssl: make ssl_sock_free_srv_ctx() zero the pointers after free
In ssl_sock_free_srv_ctx() there are some calls to free() which are not
followed by a zeroing of the pointers. For now this function is only used
during deinit but it could be used at run time in the near future, so
better secure this.

(cherry picked from commit e709e821734c306ae482e537e3ab4960046e386c)
[wt: backported to support next patch; adjusted ctx as SNI not stored
 there in 2.4 and upper; no ha_free() in 2.3]
Signed-off-by: Willy Tarreau <w@1wt.eu>