BUG/MEDIUM: spoe: Explicitly wakeup SPOE stream if waiting for more data
It is a regression introduced by the commit 1526ce4e6 ("BUG/MINOR: spoe:
Don't systematically wakeup SPOE stream in the applet handler"). It only
affects the 1.8. If a frame is not fully received, the SPOE stream is not
properly woken up when new data are received. Thus the SPOE applet is
blocked.

To fix this issue, the SPOE stream is explicitly woken up when an applet is
waiting for more data.

This fix is 1.8 specific. Thus, there is no upstream commit ID. No backport
needed.x